NLP-Project/document_parser/requirements.txt
pdfplumber==0.10.0
pytesseract==0.3.10
pillow==10.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pandas==2.0.3